Resoluçao de conflitos, conexao entre features e pequenos ajustes
